Order Create success

+++ app/src/main/java/com/example/ead_ecommerce_mobile_new/models/Order.kt
@@ -10,7 +10,7 @@ data class Product(
     val productStatus: Boolean,
     val productAvailability: Boolean,
     val productImage: String,
-    val productPrice: Double
+    val productPrice: Int
 )
 
 class Order(
